Move SoftwareType to an internal package
